+++ choco-extra/src/main/java/solver/deprecatedPropagators/PropElementV.java
@@ -74,11 +74,6 @@ public class PropElementV extends Propagator<IntVar> {
     }
 
     @Override
-    public int getPropagationConditions(int vIdx) {
-        return EventType.INT_ALL_MASK();
-    }
-
-    @Override
     public void propagate(int evtmask) throws ContradictionException {
         int n = vars.length;
         IntVar idxVar = getIndexVar();