+++ choco-extra/src/main/java/solver/deprecatedPropagators/PropGCC_AC_Cards_Fast.java
@@ -431,11 +431,6 @@ public class PropGCC_AC_Cards_Fast extends Propagator<IntVar> {
     //***********************************************************************************
 
     @Override
-    public int getPropagationConditions(int vIdx) {
-        return EventType.INT_ALL_MASK();
-    }
-
-    @Override
     public ESat isEntailed() {
         int[] f = new int[n2];
         int idx;