+++ choco-solver/src/main/java/solver/constraints/nary/alldifferent/PropAllDiffAC_Fast.java
@@ -111,11 +111,6 @@ public class PropAllDiffAC_Fast extends Propagator<IntVar> {
         SCCfinder = new StrongConnectivityFinder(digraph);
     }
 
-    @Override
-    public int getPropagationConditions(int vIdx) {
-        return EventType.INT_ALL_MASK();
-    }
-
     //***********************************************************************************
     // PROPAGATION
     //***********************************************************************************