+++ choco-solver/src/main/java/solver/constraints/nary/alldifferent/PropAllDiffAC_adaptive.java
@@ -123,11 +123,6 @@ public class PropAllDiffAC_adaptive extends Propagator<IntVar> {
         SCCfinder = new StrongConnectivityFinder(digraph);
     }
 
-    @Override
-    public int getPropagationConditions(int vIdx) {
-        return EventType.INT_ALL_MASK();
-    }
-
     //***********************************************************************************
     // PROPAGATION
     //***********************************************************************************