+++ choco-solver/src/main/java/solver/constraints/nary/circuit/PropSubCircuitSCC.java
@@ -97,11 +97,6 @@ public class PropSubCircuitSCC extends Propagator<IntVar> {
 	//***********************************************************************************
 
 	@Override
-	public int getPropagationConditions(int vIdx) {
-		return EventType.INT_ALL_MASK();
-	}
-
-	@Override
 	public ESat isEntailed() {
 		return ESat.TRUE;// redundant propagator
 	}