+++ choco-solver/src/main/java/solver/constraints/nary/lex/PropLex.java
@@ -72,11 +72,6 @@ public class PropLex extends Propagator<IntVar> {
     }
 
     @Override
-    public int getPropagationConditions(int vIdx) {
-        return EventType.INT_ALL_MASK();
-    }
-
-    @Override
     public void propagate(int evtmask) throws ContradictionException {
         if ((evtmask & EventType.FULL_PROPAGATION.mask) != 0) {
             initialize();