+++ choco-solver/src/main/java/solver/constraints/nary/nValue/AMNV_Gci_MD_R13.java
@@ -278,11 +278,6 @@ public class AMNV_Gci_MD_R13 extends Propagator<IntVar> {
 	//***********************************************************************************
 
 	@Override
-	public int getPropagationConditions(int vIdx) {
-		return EventType.INT_ALL_MASK();
-	}
-
-	@Override
 	public ESat isEntailed() {
 		return ESat.TRUE;//redundant propagator
 	}