+++ choco-solver/src/main/java/solver/constraints/set/PropAllEqual.java
@@ -99,11 +99,6 @@ public class PropAllEqual extends Propagator<SetVar> {
     //***********************************************************************************
 
     @Override
-    public int getPropagationConditions(int vIdx) {
-        return EventType.ADD_TO_KER.mask + EventType.REMOVE_FROM_ENVELOPE.mask;
-    }
-
-    @Override
     public void propagate(int evtmask) throws ContradictionException {
         if ((evtmask & EventType.FULL_PROPAGATION.mask) != 0) {
             for (int i = 0; i < n; i++) {